+++ pkg/deployment/agency/target_test.go
@@ -35,11 +35,15 @@ func Test_Target_HotBackup(t *testing.T) {
 		require.True(t, s.Agency.Arango.Target.HotBackup.Create.Exists())
 
 		t.Log(s.Agency.Arango.Target.HotBackup.Create.time.String())
+
+		require.False(t, s.Agency.ArangoDB.ArangoSync.IsSyncInProgress())
 	})
 	t.Run("Does Not Exists", func(t *testing.T) {
 		var s DumpState
 		require.NoError(t, json.Unmarshal(agencyDump39Satellite, &s))
 
 		require.False(t, s.Agency.Arango.Target.HotBackup.Create.Exists())
+
+		require.False(t, s.Agency.ArangoDB.ArangoSync.IsSyncInProgress())
 	})
 }