+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - (Refactoring) Extract kerrors package
 - (Refactoring) Extract Inspector Definitions package
 - (Bugfix) Fix PDBs Version discovery
+- (Feature) Agency ArangoSync State check
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress